pyproject.toml
[project]
name = "agile-wargame-simulator"
version = "0.1.0"
description = "Multi-agent LLM simulation engine for predicting project friction"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "litellm>=1.0",
    "chromadb>=0.5",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "fastapi>=0.110",
    "sse-starlette>=1.6",
    "uvicorn>=0.29",
    "rich>=13.0",
    "jinja2>=3.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27"]

[project.scripts]
wargame = "wargame.cli:main"

[tool.setuptools.packages.find]
include = ["wargame*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
